requirements.txt

Flask==3.0.3
gunicorn==23.0.0
gspread==6.1.4
google-auth==2.37.0
plotly==5.24.0
pandas==2.2.3
plotly==5.24.0